+++ src/test/java/com/kgu/studywithme/acceptance/category/CategoryAcceptanceFixture.java
@@ -1,11 +1,11 @@
-package com.kgu.studywithme.acceptance.fixture;
+package com.kgu.studywithme.acceptance.category;
 
 import io.restassured.response.ValidatableResponse;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 
-import static com.kgu.studywithme.acceptance.fixture.CommonRequestFixture.getRequest;
+import static com.kgu.studywithme.acceptance.CommonRequestFixture.getRequest;
 
 public class CategoryAcceptanceFixture {
     public static ValidatableResponse 모든_스터디_카테고리를_조회한다() {
