+++ src/test/java/com/kgu/studywithme/acceptance/favorite/FavoriteAcceptanceFixture.java
@@ -1,0 +1,4 @@
+package com.kgu.studywithme.acceptance.favorite;
+
+public class FavoriteAcceptanceFixture {
+}
