+++ src/test/java/com/kgu/studywithme/acceptance/member/MemberAcceptanceTest.java
@@ -1,10 +1,10 @@
-package com.kgu.studywithme.acceptance;
+package com.kgu.studywithme.acceptance.member;
 
 import com.kgu.studywithme.common.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.kgu.studywithme.acceptance.fixture.MemberAcceptanceFixture.회원가입을_진행한다;
+import static com.kgu.studywithme.acceptance.member.MemberAcceptanceFixture.회원가입을_진행한다;
 import static com.kgu.studywithme.common.fixture.MemberFixture.JIWON;
 import static org.springframework.http.HttpStatus.CREATED;
