+++ src/test/java/com/kgu/studywithme/acceptance/member/MemberQueryAcceptanceTest.java
@@ -1,4 +1,4 @@
-package com.kgu.studywithme.acceptance;
+package com.kgu.studywithme.acceptance.member;
 
 import com.kgu.studywithme.common.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
