+++ src/test/java/com/kgu/studywithme/acceptance/upload/UploadAcceptanceFixture.java
@@ -1,0 +1,4 @@
+package com.kgu.studywithme.acceptance.upload;
+
+public class UploadAcceptanceFixture {
+}
